+++ src/main/java/co/topc/iems/entity/Room.java
@@ -29,7 +29,7 @@ public class Room implements Serializable {
     /**
      * 房间面积
      */
-    private BigDecimal roomArea;
+    private Double roomArea;
 
     /**
      * 房间状态
@@ -94,11 +94,11 @@ public class Room implements Serializable {
         this.roomNo = roomNo;
     }
 
-    public BigDecimal getRoomArea() {
+    public Double getRoomArea() {
         return roomArea;
     }
 
-    public void setRoomArea(BigDecimal roomArea) {
+    public void setRoomArea(Double roomArea) {
         this.roomArea = roomArea;
     }
